Create import and export features
Layers, reports and dashboards can now be exported and imported to and
from different instances of Urungi.
The import and export pages were created in the frontend, and can be
accessed with their urls : "#/import" and "#/export"
They are not linked to within the main urungi interface
Graphic work is minimal. For now, the feature is only meant to be used
for urungi installation.